Refactor in PersistAuthorDTO. Add validation rules to avoid unnecessary calls to database.

+++ PortfolioWeb.Api/Controllers/AuthorsController.cs
@@ -26,6 +26,13 @@ public class AuthorsController(IAuthorService authorService) : ControllerBase
     [HttpGet("{id:guid}")]
     public async Task<ActionResult<AuthorDTO?>> GetById(Guid id, CancellationToken cancellationToken)
     {
+        var invalidIdResult = ValidateAuthorId(id);
+
+        if (invalidIdResult is not null)
+        {
+            return invalidIdResult;
+        }
+
         try
         {
             var author = await authorService.GetById(id, cancellationToken);
@@ -44,7 +51,7 @@ public class AuthorsController(IAuthorService authorService) : ControllerBase
     }
 
     [HttpPost]
-    public async Task<ActionResult<AuthorDTO>> Create([FromBody] CreateAuthorDTO authorDto, CancellationToken cancellationToken)
+    public async Task<ActionResult<AuthorDTO>> Create([FromBody] PersistAuthorDTO authorDto, CancellationToken cancellationToken)
     {
         try
         {
@@ -59,8 +66,15 @@ public class AuthorsController(IAuthorService authorService) : ControllerBase
     }
 
     [HttpPut("{id:guid}")]
-    public async Task<ActionResult<AuthorDTO>> Update(Guid id, [FromBody] CreateAuthorDTO authorDto, CancellationToken cancellationToken)
+    public async Task<ActionResult<AuthorDTO>> Update(Guid id, [FromBody] PersistAuthorDTO authorDto, CancellationToken cancellationToken)
     {
+        var invalidIdResult = ValidateAuthorId(id);
+
+        if (invalidIdResult is not null)
+        {
+            return invalidIdResult;
+        }
+
         try
         {
             var updatedAuthor = await authorService.Update(id, authorDto, cancellationToken);
@@ -81,6 +95,13 @@ public class AuthorsController(IAuthorService authorService) : ControllerBase
     [HttpDelete("{id:guid}")]
     public async Task<IActionResult> Delete(Guid id, CancellationToken cancellationToken)
     {
+        var invalidIdResult = ValidateAuthorId(id);
+
+        if (invalidIdResult is not null)
+        {
+            return invalidIdResult;
+        }
+
         try
         {
             var wasDeleted = await authorService.Delete(id, cancellationToken);
@@ -97,4 +118,6 @@ public class AuthorsController(IAuthorService authorService) : ControllerBase
             return StatusCode(StatusCodes.Status500InternalServerError);
         }
     }
+
+    private ActionResult? ValidateAuthorId(Guid id) => id == Guid.Empty ? BadRequest("The provided author id is not valid.") : null;
 }
